chore: Some bits here and there

+++ client/src/main/java/com/jagex/runetek4/Static40.java
@@ -26,7 +26,7 @@ public final class Static40 {
 
 	@OriginalMember(owner = "client!da", name = "c", descriptor = "(II)Lclient!aa;")
 	public static MSIType get(@OriginalArg(0) int arg0) {
-		@Pc(10) MSIType local10 = (MSIType) PreciseSleep.aClass99_29.get((long) arg0);
+		@Pc(10) MSIType local10 = (MSIType) PreciseSleep.aClass99_29.get(arg0);
 		if (local10 != null) {
 			return local10;
 		}
@@ -35,7 +35,7 @@ public final class Static40 {
 		if (bytes != null) {
 			local10.decode(new Packet(bytes));
 		}
-		PreciseSleep.aClass99_29.put(local10, (long) arg0);
+		PreciseSleep.aClass99_29.put(local10, arg0);
 		return local10;
 	}
